fix: auth user routes
