--- frontend/src/pages/Dashboard/index.ts
@@ -1,1 +1,0 @@
-export { Dashboard } from './Dashboard';
