+++ frontend/src/router/authUserRoutes.tsx
@@ -1,10 +1,10 @@
-import { Main, NotFound, Profile } from '../pages';
+import { MainPage, NotFound, Profile } from '../pages';
 
 import { TypedNavigate } from './components/TypedNavigate';
 
 export const authUserRoutes = [
     { path: '/sign-in', element: <TypedNavigate to="/" replace /> },
-    { path: '/', element: <Main /> },
+    { path: '/', element: <MainPage /> },
     { path: '/profile', element: <Profile /> },
     { path: '/admin', element: <div>Admin page</div> },
     { path: '/rules', element: <div>Program rules page</div> },
