+++ frontend/src/utils/range.ts
@@ -1,0 +1,7 @@
+export function range(count: number): number[] {
+    const r: number[] = [];
+    for (let i = 0; i < count; i++) {
+        r.push(i);
+    }
+    return r;
+}
